+++ features/Changelog.md
@@ -2,7 +2,8 @@
 
 [full changelog](http://github.com/rspec/rspec-expectations/compare/v2.5.0...v2.6.0)
 
-* Enhancments
+* Enhancements
+  * HaveMatcher converts argument using #to_i (Alex Bepple & Pat Maddox)
   * `change` matcher accepts Regexps (Robert Davis)
   * better descriptions for have_xxx matchers (Magnus Bergmark)
   * range.should cover(*values) (Anders Furseth)